+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # PhotoGallery Application
 
+# 0.0.2
+**Added**
+	- Identity db context and initial models
+	- PackageModel
+	- ApplicationUser
+
+**Changed**
+	- ApplicationUser extends IdentityUser
+	- PackagePlan models part of Identity db context
+	- Profile page extended
+	- Navigation updated
+
 # 0.0.1
-- **Added**
- - initial projects structure and nugets
- - initial model for packages plans
+**Added**
+	- initial projects structure and nugets
+	- initial model for packages plans